x86/up: fix datatypes for >32-bit CPLD registers
Some lines weren't updated to support CPLD registers over 32bits.
This prevented the LEDs from working on DVT.

+++ drivers/platform/x86/up_board_pinctrl.c
@@ -544,12 +544,12 @@ static int cpld_configure(struct up_cpld_info *cpld)
 static int cpld_set_value(struct up_cpld_info *cpld, unsigned int offset,
 			  int value)
 {
-	u32 old_regval = cpld->dir_reg;
+	u64 old_regval = cpld->dir_reg;
 
 	if (value)
-		cpld->dir_reg |= 1 << offset;
+		cpld->dir_reg |= 1ULL << offset;
 	else
-		cpld->dir_reg &= ~(1 << offset);
+		cpld->dir_reg &= ~(1ULL << offset);
 
 	if (cpld->dir_reg != old_regval)
 		return cpld_configure(cpld);
